// test/packages/DocumentManage/src/pages/_dto/model/DocumentModel.ts
export interface DocumentModel {
  appNo: string;
  batchNo: string;
  businessNo: string;
  caseCategory: string;
  caseNo: string;
  classification: number;
  content: string;
  contentType: number;
  creationDate: Date;
  creator: string;
  deleted: number;
  docId: string;
  docTypeCode: string;
  formCategory: string;
  formerDocId: string;
  gmtCreate: Date;
  gmtModified: Date;
  id: string;
  image: string;
  indexClass: string;
  insuredName: string;
  modifier: string;
  name: string;
  new: true;
  parentBusinessNo: string;
  policies: string;
  receivedDate: Date;
  replaceDocId: string;
  srcDocId: string;
  submissionId: string;
  transId: string;
  type: number;
  voidFlag: number;
  imageUploadStatus: string;
  customerType: string;
  customerName: string;
  mustRead: boolean;
  unRead: boolean;
}